IGNITE-35 Fixed test

+++ modules/core/src/test/java/org/gridgain/grid/p2p/GridP2PDifferentClassLoaderSelfTest.java
@@ -95,8 +95,8 @@ public class GridP2PDifferentClassLoaderSelfTest extends GridCommonAbstractTest 
             }
 
             // Execute task1 and task2 from node1 on node2 and make sure that they reuse same class loader on node2.
-            Integer res1 = ignite1.compute().execute(task1, ignite2.cluster().localNode().id());
-            Integer res2 = ignite1.compute().execute(task2, ignite2.cluster().localNode().id());
+            Integer res1 = (Integer)ignite1.compute().execute(task1, ignite2.cluster().localNode().id());
+            Integer res2 = (Integer)ignite1.compute().execute(task2, ignite2.cluster().localNode().id());
 
             if (expectEquals)
                 assert res1.equals(res2);
